requirements.txt
elasticsearch>=8.0.0,<9.0.0
elastic-transport>=8.0.0,<9.0.0
python-dotenv
openai>=1.0.0 
pytest>=7.4.0
numpy
pandas

pdfplumber>=0.10.0
pdfminer.six>=20221105
PyYAML>=6.0
google-auth>=2.17.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.88.0
pydantic>=2.0.0
click>=8.0.0
jsonschema>=4.17.0
markdownify
requests>=2.31.0
beautifulsoup4>=4.12.0
chardet>=5.0.0

# OCR dependencies for image-based PDFs (optional)
pytesseract
Pillow
PyMuPDF

# Async HTTP client for spreadsheet processing
aiohttp>=3.8.0
